[Altimetry] do not throw exception for invalid altitudes (related to #1472863)
When the AltimetryBuilder can not calculate an altitude it warns but does not
throws. Throwing here puts too much constraints on the client code.

Instead the client code is expected to check and handle altitude validity in an
explicit way. The throwing of the exception is moved when exporting to a solver
data model points with invalid altitude.