[TYRoute] Simplyfied model wrt. to road declivity
closes #1513437

A road is modelized by a single lineic source but the NMPB2008 model
emission model take declivity into account which would need to
represent the two opposite roadways.

As a simplified model we split the traffic equally between the two
roadways and sum the emission spectrum obtained for opposite
declivity.