[SolverDataModelBuilder transfer] Rewrite processAltimetry CAUTION
For now, wrap a pointer to an AcousticProblemModel in the python object
ProblemModel and make the ModelBuilder construct and return it on demand
because dealing with references in cython appears problematical.

Declare temporarily SolverDataModelBuilder::processMesh as public to be able
to call it from the new process_altimetry function. This call will be replaced
by a cython implementation in the next changeset and reestablished as protected.

CAUTION test "test_base" KO. This test is reestablished in cset
"[SolverdataModelBuilder transfer] Stop using the SolverDataModelBuilder"

related to #1587978