[cppcheck] drop never used variables
Messages below have not been considered and a deeper look at them would be welcome

[Tympan/Appli/TympanApp/TYSilosEditor.cpp:197]: (style) Variable 'pAction' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSilosEditor.cpp:241]: (style) Variable 'pAction' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYElementPicker.cpp:112]: (style) Variable 'z2' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYPickEditor.cpp:456]: (style) Variable 'pEltGeoNode' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYProjetFrame.cpp:161]: (style) Variable
'pPointControlItem' is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYProjetFrame.cpp:193]: (style) Variable
'pAtmosphereItem' is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYProjetFrame.cpp:194]: (style) Variable 'pEtatItem' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYProjetFrame.cpp:200]: (style) Variable 'pMaillageItem'
is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYProjetFrame.cpp:203]: (style) Variable 'pResultatItem'
is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYProjetFrame.cpp:271]: (style) Variable
'pPopupMenuSpectre' is allocated memory that is never used.
[Tympan/Appli/TympanApp/TYProjetFrame.cpp:272]: (style) Variable
'pPopupMenuLinearMaillage' is allocated memory that is never used.
[Tympan/Appli/TympanApp/TYProjetFrame.cpp:273]: (style) Variable
'pPopupMenuRectangularMaillage' is allocated memory that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:191]: (style) Variable 'pPlanEauItem' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:199]: (style) Variable 'pCoursEauItem'
is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:207]: (style) Variable 'pCourbeNivItem'
is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:243]: (style) Variable 'pSourceItem' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:260]: (style) Variable 'pMachineItem' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:268]: (style) Variable 'pResTranspItem'
is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:276]: (style) Variable 'pRouteItem' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:287]: (style) Variable
'pSpectreTransmItem' is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:288]: (style) Variable
'pSpectreAbsoItem' is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:300]: (style) Variable 'pVegetationItem'
is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:333]: (style) Variable 'pAccVolItem' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:336]: (style) Variable
'pListSpectreItem' is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:340]: (style) Variable
'pAttenuateurItem' is assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:375]: (style) Variable 'pMachineItem' is
assigned a value that is never used.
[Tympan/Appli/TympanApp/TYSiteFrame.cpp:382]: (style) Variable 'pSourceItem' is
assigned a value that is never used.
[Tympan/MetierSolver/AcousticRaytracer/Accelerator/GridAccelerator.cpp:144]:
(style) Variable 'hitSomething' is assigned a value that is never used.
[Tympan/MetierSolver/AcousticRaytracer/Engine/DefaultEngine.cpp:201]: (style)
Variable 'find_intersection' is assigned a value that is never used.
[Tympan/MetierSolver/DataManagerMetier/Commun/TYCalcul.cpp:1163]: (style)
Variable 'cancel' is assigned a value that is never used.
[Tympan/MetierSolver/DataManagerMetier/ComposantGeoAcoustique/TYAcousticLine.cpp:396]:
(style) Variable 's' is assigned a value that is never used.
[Tympan/MetierSolver/DataManagerMetier/Parcours/TYCalculParcours.cpp:321]:
(style) Variable 'bPolylignesInfraFermees' is assigned a value that is never
used.
[Tympan/MetierSolver/SolverDataModel/acoustic_problem_model.cpp:26]: (warning)
Member variable 'AcousticTriangle::uuid' is not initialized in the constructor.
[Tympan/MetierSolver/SolverDataModel/data_model_common.hpp:75]: (style) struct
or union member 'Anonymous0::data1' is never used.
[Tympan/MetierSolver/SolverDataModel/data_model_common.hpp:76]: (style) struct
or union member 'Anonymous0::data2' is never used.
[Tympan/MetierSolver/SolverDataModel/data_model_common.hpp:77]: (style) struct
or union member 'Anonymous0::data3' is never used.
[Tympan/MetierSolver/SolverDataModel/data_model_common.hpp:78]: (style) struct
or union member 'Anonymous0::data4' is never used.
[tests/business_logic/common/test_TYRay.cpp:32]: (style) Variable 'type_ev' is
assigned a value that is never used.
[tests/business_logic/infrastructure/test_road.cpp:37]: (style) Variable 'xMin'
is assigned a value that is never used.
[tests/business_logic/infrastructure/test_road.cpp:37]: (style) Variable 'xMax'
is assigned a value that is never used.
[tests/business_logic/infrastructure/test_road.cpp:37]: (style) Variable 'yMin'
is assigned a value that is never used.
[tests/business_logic/infrastructure/test_road.cpp:37]: (style) Variable 'yMax'
is assigned a value that is never used.
[tests/business_logic/topography/test_TYAltimetrie.cpp:100]: (style) Variable
'xMin' is assigned a value that is never used.
[tests/business_logic/topography/test_TYAltimetrie.cpp:100]: (style) Variable
'xMax' is assigned a value that is never used.
[tests/business_logic/topography/test_TYAltimetrie.cpp:100]: (style) Variable
'yMin' is assigned a value that is never used.
[tests/business_logic/topography/test_TYAltimetrie.cpp:100]: (style) Variable
'yMax' is assigned a value that is never used.