[tests] Add datasets to setup functional tests
Empty projects (IE acoustic problems without results) are the
TEST_XXX_NO_RESU.xml files, added to the "projects-panel" folder.

The corresponding projets with the expected results are the TEST_XXX.xml
files,
added to the "expected" folder.

Caution: the <useEcran> value has been set to 1 in all the projects-panel
files so there is a slight difference with the files originating from EDF.


Related to #1607848