Allow to run Tympan using legacy computation (no python)
Add a basic command-line argument handling for Tympan in
TYApplication: when --lecagy-compuation is given, the new
data-path through python is bypassed.

CAUTION This option will NOT be maintained and is to be considered
as forever experimental ! It WILL be dropped when the new
data-path is validated, or even before it need be.