Add a `build_model` method to Project
This method builds the solver model. For now, it also returns the converter
instance.

Drop nsources and nreceptors properties and related attributes from
Business2SolverConverter and use respective accessors from the solver model.
Yet add assertions to ensure these are consistent with the returns of
build_sources and build_receptors methods of SolverModelBuilder.

Related to #2798352.