[solvers] allow to choose the computation solver from the python side
The solvers have to be loaded before we read there information (with 'solver_id()' for example).

That is why split "load_solver()" C++ function into a function in charge of loading the solver, and
another one in charge of selecting a solver.
This way the "load_solvers()" function can be reused.