Another approach to multi-segments screens handling
This reverts c3a0c53fcad0 by and large and should fix bitbucket issue #3.

In this new approach, we keep all faces on ground of infrastructure elements,
thus TYSiteNode::groundBasedFaces now returns a map of UUID to a deque of
arrays of 3D points (instead of mapping to a single array of 3D points).
These faces are then passed on to the altimetry processing. When building the
altimetry model from Tympan infrastructure elements, we also keep all polygons
on ground so that tympan.altimetry.datamodel.InfrastructureLandtake is now
built using several polygons (adding a dedicated test with an infrastructure
element defined with several polygons in test_altimetry_builder.py).
Accordingly, when building the mesh, we need to build groups of vertices
(corresponding to each polygon) for each geometric features. This new aspect
is handled later on when filling the mesh with material information.